+++ src/main/scala/com/raquo/airstream/features/CombineObservable.scala
@@ -13,8 +13,8 @@ trait CombineObservable[A] extends SyncObservable[A] { self =>
   /** Parent observers are not immediately active. onStart/onStop regulates that. */
   protected[this] val parentObservers: js.Array[InternalParentObserver[_]] = js.Array()
 
-  // Implementations should call internalObserver.onNext() instead of .fire()
-  // Transaction will call .fireSync() when it's time, and that will in turn call .fire()
+  // Implementations should call internalObserver.onNext() instead of .fireValue()
+  // Transaction will call .fireSync() when it's time, and that will in turn call .fireValue()
   protected[this] val internalObserver: InternalObserver[A] = InternalObserver.fromTry[A]((nextValue, transaction) => {
     if (!transaction.pendingObservables.contains(self)) {
       // println(s"Marking CombineObs($id) as pending in TRX(${transaction.id})")