+++ src/main/scala/com/raquo/airstream/signal/Tuple2Signal.scala
@@ -1,5 +1,7 @@
 package com.raquo.airstream.signal
 
+// @TODO[API] I think this can be a Tuple2Observable, or maybe even an actual native method with evidence
+
 class Tuple2Signal[A, B](val tuple2Signal: Signal[(A, B)]) extends AnyVal {
 
   def map2[C](project: (A, B) => C): Signal[C] = {